all basic functions implemented

+++ FlooringMastery/src/main/java/com/sg/flooringmastery/service/ServiceLayerImpl.java
@@ -9,14 +9,14 @@ import java.time.LocalDate;
 import java.util.List;
 import java.util.Map;
 
-public class FlooringMasteryServiceLayerImpl implements FlooringMasteryServiceLayer{
+public class ServiceLayerImpl implements ServiceLayer {
 
     private TaxDao taxDao;
     private ProductDao productDao;
     private OrderDao orderDao;
     private String mode;
 
-    public FlooringMasteryServiceLayerImpl(TaxDao taxDao, ProductDao productDao, OrderDao orderDao) {
+    public ServiceLayerImpl(TaxDao taxDao, ProductDao productDao, OrderDao orderDao) {
         this.orderDao = orderDao;
         this.productDao = productDao;
         this.taxDao = taxDao;
@@ -28,36 +28,53 @@ public class FlooringMasteryServiceLayerImpl implements FlooringMasteryServiceLa
     }
 
     @Override
-    public Order addOrder(Order orderToAdd) throws Exception{
+    public Order addOrder(Order orderToAdd) throws OrderPersistenceException{
         return orderDao.createOrder(orderToAdd);
     }
 
     @Override
-    public void saveCurrentWork(Map<Integer, Order> orderMap) throws OrderPersistenceException {
+    public void saveCurrentWork() throws OrderPersistenceException {
         orderDao.save();
     }
 
     @Override
     public Order updateOrder(Order orderToUpdate)throws OrderPersistenceException {
-        return orderDao.updateOrder(orderToUpdate);
+        if (validateOrderId(orderToUpdate.getOrderDate(), orderToUpdate.getOrderNumber())==true) {
+            return orderDao.updateOrder(orderToUpdate);
+        } else throw new OrderPersistenceException("ERROR: Order does not exist.") ;
+
     }
 
     @Override
     public void removeOrder(Order orderToRemove) throws OrderPersistenceException {
-        orderDao.removeOrder(orderToRemove);
+        if (validateOrderId(orderToRemove.getOrderDate(), orderToRemove.getOrderNumber())){
+            orderDao.removeOrder(orderToRemove);
+        } else throw new OrderPersistenceException("ERROR: Order does not exist.");
     }
 
     @Override
     public Order retrieveOrderByDateAndId(LocalDate date, int orderId) throws OrderPersistenceException{
-        return orderDao.retrieveOrderByDateAndId(date, orderId);
+        if(validateOrderId(date, orderId)) {
+            return orderDao.retrieveOrderByDateAndId(date, orderId);
+        } else throw new OrderPersistenceException("ERROR: Order does not exist.");
     }
 
     @Override
-    public Order processOrder(Order orderToProcess) throws TaxPersistenceException, ProductPersistenceException {
+    public Order processOrder(Order orderToProcess) throws TaxPersistenceException, ProductPersistenceException,
+            OrderPersistenceException{
         //Call the 4 calculate methods to set the 4 calculated fields.
         //Fields that will be user entered: customer lastname, state, product type, area.
         //Fields to be populated by this method: Tax rate, material cost per square foot,
         //labor cost per square foot, plus the 4 calculated fields.
+        //TODO: implement negative tests to verify the exception conditions.
+
+        if(!validateProductExists(orderToProcess)) {
+            throw new ProductPersistenceException("ERROR: Product does not exist.");
+        }
+
+        if(!validateStateExists(orderToProcess)) {
+            throw new TaxPersistenceException("ERROR: State does not exist.");
+        }
 
         orderToProcess.setOrderTax(taxDao.retrieveTax(orderToProcess.getState()));
         orderToProcess.setOrderProduct(productDao.retrieveProduct(orderToProcess.getOrderProduct().getProductType()));
@@ -100,10 +117,10 @@ public class FlooringMasteryServiceLayerImpl implements FlooringMasteryServiceLa
         } else return true;
     }
 
-    private boolean validateOrderExists(Order order) throws OrderPersistenceException {
-        if (orderDao.retrieveOrderByDateAndId(order.getOrderDate(),
-                order.getOrderNumber()) == null) {
+    private boolean validateOrderId(LocalDate localDate, int orderId) throws OrderPersistenceException {
+        if (orderDao.retrieveOrderByDateAndId(localDate, orderId) == null) {
+
             return false;
-        } else return true;
+        } return true;
     }
 }
